64-bit fixes
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@17638 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/willow/whttp_header.cc
@@ -220,7 +220,7 @@ size_t	nbufsz = curnlen + curvlen + 4 + len;
 	char	*nbuf = hl_last->alloc.allocate(nbufsz);
 		strcpy(nbuf, hl_last->hr_name);
 		sprintf(nbuf + curnlen + 1, "%s, %.*s",
-			hl_last->hr_value, len, append);
+			hl_last->hr_value, (int) len, append);
 
 		if (hl_last->hr_allocd)
 			hl_last->alloc.deallocate(hl_last->hr_name, hl_last->hr_allocd);
